Users can now logout and added user restriction so that only room hosts can edit or delete their rooms

+++ base/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic, User
 from .forms import RoomForm
@@ -18,12 +19,23 @@ def login_view(request):
 
         user = authenticate(request, username=username, password=password)
 
-        if user is not None:
+        if user is None:
+            messages.error(request, 'Incorrect password.')
+        else:
             login(request, user)
             return redirect('home')
 
+    if request.user.is_authenticated:
+        return redirect('home')
+        
     context = {}
     return render(request, 'base/login_register.html', context)
+
+
+@login_required(login_url='login')
+def logout_view(request):
+    logout(request)
+    return redirect('home')
 
 
 def home(request):
@@ -49,6 +61,7 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
      
 
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
     
@@ -62,9 +75,14 @@ def create_room(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if room.host != request.user:
+        messages.error(request, 'You do not own this room.')
+        return redirect('home')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -76,8 +94,14 @@ def update_room(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if room.host != request.user:
+        messages.error(request, 'You do not own this room.')
+        return redirect('home')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
